Changed installer to use jar for path editing
The PathEd.exe tool is great, except that it's dependent on .NET. The built-in NSIS tools for string manipulation are terrible, leading to either complicated code for path editing or a high chance of errors.

This PathEditor.jar (https://github.com/bwittman/PathEditor) is a port of the PathEd (https://github.com/awaescher/PathEd) tool to Java. Since Java is required for the Shadow compiler to run, it's a better requirement than a lot of .NET stuff.